Desktop/retail_etl_project/requirements.txt
apache-airflow
apache-airflow-core
apache-airflow-providers-amazon==9.9.0
apache-airflow-providers-common-compat==1.7.1
apache-airflow-providers-common-io==1.6.0
apache-airflow-providers-common-sql==1.27.2
apache-airflow-providers-http==5.3.1
apache-airflow-providers-smtp==2.1.0
apache-airflow-providers-snowflake==6.4.0
apache-airflow-providers-standard==1.3.0
apache-airflow-task-sdk==1.0.2
boto3==1.38.44
botocore==1.38.44
jsonpath-ng==1.7.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
numpy==1.26.4
pandas==2.1.4
pandera==0.24.0
pendulum==3.1.0
pyarrow==18.0.0
PyYAML==6.0.2
snowflake-connector-python==3.16.0
snowflake-snowpark-python==1.35.0
snowflake-sqlalchemy==1.7.6
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
